requirements.txt
Django==4.0.5
pyodbc==4.0.39
django-mssql==1.8
django-pyodbc-azure==1.2.5
djangorestframework==3.14.0
asgiref==3.6.0
mssql==1.0.1
mssql-django==1.1.3
pytz==2022.7
sqlparse==0.4.3
tzdata==2022.7